runtime: move unsafe code to a separate module.
